+++ app/src/main/java/com/katiearose/sobriety/AddictionCardAdapter.kt
@@ -6,6 +6,8 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.katiearose.sobriety.activities.Main
+import com.katiearose.sobriety.internal.CacheHandler
 
 class AddictionCardAdapter(private val activity: Main, private val cacheHandler: CacheHandler) :
     RecyclerView.Adapter<AddictionCardAdapter.AddictionCardViewHolder>() {
