+++ app/src/main/java/com/katiearose/sobriety/internal/CircularBuffer.kt
@@ -1,4 +1,4 @@
-package com.katiearose.sobriety
+package com.katiearose.sobriety.internal
 
 import java.io.Serializable
