+++ app/src/main/res/menu/menu_main.xml
@@ -1,2 +1,2 @@
 <menu xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.katiearose.sobriety.Main" />
+    tools:context="com.katiearose.sobriety.activities.Main" />
